Update document source flag types
